+++ CHANGELOG.md
@@ -10,6 +10,7 @@ described with adjustments and template in [TSWiki](https://tswiki.corp.saab.se/
 ### Added
 - Support for setting Unity layer for Node Builders 
 - Support for setting Unity layer for Foliage 
+- New WorldToUnity methods to MapUtil
 
 ## 4.3